fixed up file rename and removed a console.log()
