+++ samples/RefreshFlow/AngularApp/src/core/auth-store/auth.store.ts
@@ -1,10 +1,10 @@
 import { ActionReducer, combineReducers } from '@ngrx/store';
 import { ProfileModel } from '../models/profile-model';
-import { AuthTokenModel } from '../models/auth-tokens.model';
 import { profileReducer } from '../profile/profile.reducer';
 import { authTokenReducer } from '../auth-token/auth-token.reducer';
 import { authReadyReducer } from './auth-ready.reducer';
 import { loggedInReducer } from './logged-in.reducer';
+import { AuthTokenModel } from '../models/auth-tokens-model';
 
 export interface AuthState {
     authTokens: AuthTokenModel;
