+++ samples/RefreshFlow/AngularApp/src/core/auth-token/auth-token.actions.ts
@@ -1,7 +1,7 @@
 import { Action } from '@ngrx/store';
 import { type } from '../../util/action-name-helper';
-import { AuthTokenModel } from '../models/auth-tokens.model';
 import { Injectable } from '@angular/core';
+import { AuthTokenModel } from '../models/auth-tokens-model';
 
 export const AuthTokenActionTypes = {
     LOAD: type('[AuthToken] Load'),
